changed forward datatype to double() in flows/transforms/made.py

+++ nflows/transforms/made.py
@@ -69,7 +69,7 @@ class MaskedLinear(nn.Linear):
         return mask, out_degrees
 
     def forward(self, x):
-        return F.linear(x, self.weight * self.mask, self.bias)
+        return F.linear(x.double(), self.weight * self.mask, self.bias)
 
 
 class MaskedFeedforwardBlock(nn.Module):
